requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Mail==0.10.0
Flask-WTF==1.2.2
gcloud==0.18.3
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jws==0.1.3
MarkupSafe==3.0.2
oauth2client==4.1.3
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pyparsing==3.2.1
pypdfium2==4.30.1
python-dotenv==1.0.1
python-jwt==2.0.1
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
setuptools==75.8.0
six==1.17.0
stripe==11.6.0
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.1.3
WTForms==3.2.1
